Fixed alt attribute appearing in <object> tag for SVG images

+++ testing/runtests.py
@@ -331,7 +331,7 @@ class TestManager:
 			tester = Tester(self.args,test)
 			tester.run(self)
 		res=self.result()
-		if self.args.xhtml and not res and not self.args.keep:
+		if self.args.xhtml and self.args.inputdir!='.' and not res and not self.args.keep:
 			shutil.rmtree("dtd",ignore_errors=True)
 		return 0 if self.args.updateref else res
